requirements.txt
alembic==1.11.1
APScheduler==3.10.1
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
MLB-StatsAPI==1.6.1
passlib==1.7.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.17
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
